@@ -6,7 +6,7 @@ cmake_minimum_required (VERSION 3.8)
 project (ecs CXX)
 
 # Require c++17
-set(CMAKE_CXX_STANDARD 17)
+set(CMAKE_CXX_STANDARD 20)
 set(CXX_STANDARD_REQUIRED on)
 
 # Include google benchmark
