@@ -13,7 +13,7 @@ size_t constexpr end_range = 16 * 1024 * 1024;
 //size_t constexpr end_range = 1024 * 1024 * 1024;  // 1 billion entities
 
 struct shared_s {
-	ecs_flags(ecs::shared);
+	ecs_flags(ecs::share);
 	size_t dimension;
 };
 
