@@ -11,7 +11,7 @@ namespace ecs
 	// meaning that any entity with a shared component will all point to the same component.
 	// Think of it as a static member variable in a regular class.
 	// Uses O(1) memory instead of O(n).
-	struct shared {};
+	struct share {};
 
 	// Add this in 'ecs_flags()' to mark a component as transient.
 	// The component will only exist on an entity for one cycle,
@@ -25,29 +25,17 @@ namespace ecs
 
 	// Add flags to a component to change its behaviour and memory usage.
 	// Example:
-	// struct my_component {
+	// struct my_component { 
 	// 	ecs_flags(ecs::tag, ecs::transient);
 	// 	// component data
 	// };
 	#define ecs_flags(...) struct _ecs_flags : __VA_ARGS__ {};
 
 	namespace detail {
-		template <typename T> constexpr auto is_tagged(...)    -> bool { return false; }
-		template <typename T> constexpr auto is_shared(...)    -> bool { return false; }
-		template <typename T> constexpr auto is_transient(...) -> bool { return false; }
-		template <typename T> constexpr auto is_immutable(...) -> bool { return false; }
-
-		template <typename T> constexpr auto is_tagged(int)    -> decltype(typename T::_ecs_flags{}, bool{}) { return std::is_base_of_v<ecs::tag, typename T::_ecs_flags>; }
-		template <typename T> constexpr auto is_shared(int)    -> decltype(typename T::_ecs_flags{}, bool{}) { return std::is_base_of_v<ecs::shared, typename T::_ecs_flags>; }
-		template <typename T> constexpr auto is_transient(int) -> decltype(typename T::_ecs_flags{}, bool{}) { return std::is_base_of_v<ecs::transient, typename T::_ecs_flags>; }
-		template <typename T> constexpr auto is_immutable(int) -> decltype(typename T::_ecs_flags{}, bool{}) { return std::is_base_of_v<ecs::immutable, typename T::_ecs_flags>; }
-
 		// Some helpers
-		template <class T> constexpr static bool is_tagged_v = is_tagged<T>(0);
-		template <class T> constexpr static bool is_shared_v = is_shared<T>(0);
-		template <class T> constexpr static bool is_transient_v = is_transient<T>(0);
-		template <class T> constexpr static bool is_immutable_v = is_immutable<T>(0);
-
-		template <class T> constexpr static bool is_static_v = (is_shared_v<T> || is_tagged_v<T>); // all entities point to the same component
+		template<typename T> concept tagged = std::is_base_of_v<ecs::tag, typename T::_ecs_flags>;
+		template<typename T> concept shared = std::is_base_of_v<ecs::share, typename T::_ecs_flags>;
+		template<typename T> concept transient = std::is_base_of_v<ecs::transient, typename T::_ecs_flags>;
+		template<typename T> concept immutable = std::is_base_of_v<ecs::immutable, typename T::_ecs_flags>;
 	}
 }
