@@ -2,13 +2,13 @@
 #include "catch.hpp"
 
 struct _dummy { ecs_flags(ecs::transient); };
-static_assert(ecs::detail::is_transient_v<_dummy>);
+static_assert(ecs::detail::transient<_dummy>);
 
 TEST_CASE("Transient components", "[component][transient]")
 {
 	struct foo {};
 	struct test_t { ecs_flags(ecs::transient); };
-	static_assert(ecs::detail::is_transient_v<test_t>);
+	static_assert(ecs::detail::transient<test_t>);
 
 	ecs::detail::_context.reset();
 
